Bump to v0.3.0, sandbox by stage, universal paths, psql on slaves.
We sandbox installation directories by stage, so we can stream
production server's slave to staging server without conflicting with
staging server's master.
Tested with PostgreSQL versions 9.1, 9.2, 9.3, 9.4, and 9.5, - we remove
version specific paths, and add a setting which sets this number. We
also add a task to enter any of the slave instances with psql.